requirements.txt
Flask==3.0.0
groq==0.13.0
requests==2.31.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
